Fixed some more typos and smaller errors.

+++ input/MouseHandler.cpp
@@ -9,8 +9,8 @@
 namespace voxie {
     namespace internal {
 
-        float lastX = 0;
-        float lastY = 0;
+        double lastX = 0;
+        double lastY = 0;
         bool firstMouse = true;
         bool mouseLock = false;
 
@@ -18,8 +18,8 @@ namespace voxie {
 
     void MouseHandler::framebuffer_size_callback(GLFWwindow*, int width, int height) {
         glViewport(0, 0, width, height);
-        voxie::Engine::GetEngine().GetWindow()->SetWidth(width);
-        voxie::Engine::GetEngine().GetWindow()->SetHeight(height);
+        voxie::Engine::GetEngine().GetWindow()->SetWidth(static_cast<unsigned int>(width));
+        voxie::Engine::GetEngine().GetWindow()->SetHeight(static_cast<unsigned int>(height));
     }
 
     void MouseHandler::mouse_callback(GLFWwindow*, double xpos, double ypos) {
@@ -30,13 +30,13 @@ namespace voxie {
                 internal::firstMouse = false;
             }
 
-            float xoffset = xpos - internal::lastX;
-            float yoffset = internal::lastY - ypos;// reversed since y-coordinates go from bottom to top
+            double xoffset = xpos - internal::lastX;
+            double yoffset = internal::lastY - ypos;// reversed since y-coordinates go from bottom to top
 
             internal::lastX = xpos;
             internal::lastY = ypos;
 
-            Engine::GetEngine().GetCamera()->ProcessMouseMovement(xoffset, yoffset);
+            Engine::GetEngine().GetCamera()->ProcessMouseMovement(static_cast<float>(xoffset), static_cast<float>(yoffset));
         }
     }
